compare_evemu: ignore slots messages
two different sequences may send the same events but through different
slots. We can ignore them because if the most common case lies when
slots are swapped during the entire processing of the events.
If the date moved from one slot to the other, this case is not covered,
but it would require a kind of finger tracking for each slot.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>