testsuite.py: implement "fast mode"
The fast mode can be used to reduce the amount of tests when running
several times the tests with the same kernel release.
If there is an expected event file with the same kernel major/minor
than the one we are currently running, this means that this event file
has been created with the same kernel, meaning we can skip it.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>