compare_evemu: better handling of key repeats
key repeats generated by the kernel are marked with a value of 2. As the
kernel generate them, there is no guarantee that they will appear in the
same frame at each run.

Also, the kernel generates EV_SYN(1) to send events not triggered by the
device itself. Adding it to all frame to remove entropy during comparisons.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>