refactor the hid database and expected evemu
The current implementation implies that each test does its own
analysis to know if it should be run or skipped, and against which
evemu dumps it should compare. This does not allow a rapid check of the
database, and makes the total count number change over time.

This huge patch first creates the database of the association of the hid
devices and their corresponding evemu dumps, then launch the different
tests if they are not skipped.